libc: arm: a15: add optimized memcpy implementation
This optimization is extracted from cortex-strings and bionic-ized, then
applied to Cortex-A15 cpu. This optimization requires NEON for armv7-a or
vfpv2 for armv6. NEON is optional in cortex-a9, so I did not change
cortex-a9.mk.

This new implementation does not have __memcpy_chk(), so __memcpy_chk.cpp
is included for arch which uses the new implementation.

Change-Id: If3e11f9df4694dee3d3d8a5297817d1bc2a41d6e
Signed-off-by: Yingshiuan Pan <yingshiuan.pan@linaro.org>
Signed-off-by: arter97 <qkrwngud825@gmail.com>

Conflicts:
	libc/arch-arm/cortex-a15/cortex-a15.mk